+++ packages/db/prisma/migrations/20260202223510_add_header_image/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "AnimeSeries" ADD COLUMN     "headerImage" TEXT;